requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0
pyyaml>=6.0

# YouTube and transcript processing
pytube>=15.0.0
youtube-transcript-api>=0.6.0

# AI and language processing
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.1.0
openai>=1.0.0
faiss-cpu>=1.7.4

# Text processing
tiktoken>=0.5.0

# Export functionality
reportlab>=4.0.0

# Utilities
requests>=2.31.0
urllib3>=2.0.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
